Docs for connection, refactor for modularity, vertex
